+++ betty/plugins/entity/person.py
@@ -17,10 +17,10 @@ from betty.entity.has_notes import HasNotes
 from betty.entity.schema import ToManySchema
 from betty.functools import unique
 from betty.gender import GenderDefinition
+from betty.genders.unknown import Unknown as UnknownGender
 from betty.linked_data import JsonLdObject, dump_context
 from betty.locale.localizable.gettext import _, ngettext
 from betty.plugin.schema import PluginIdSchema
-from betty.plugins.gender.unknown import Unknown as UnknownGender
 from betty.privacy import Privacy
 
 if TYPE_CHECKING:
