+++ betty/plugins/loader/demo/jobs.py
@@ -9,6 +9,8 @@ from typing import TYPE_CHECKING, override
 
 from betty.date import Date, DateRange
 from betty.dirs import BUILTIN_ASSET_DIRECTORY
+from betty.genders.man import Man
+from betty.genders.woman import Woman
 from betty.job import Job
 from betty.locale.localizable.gettext import _
 from betty.plugins.entity.citation import Citation
@@ -27,8 +29,6 @@ from betty.plugins.entity.source import Source
 from betty.plugins.event_type.birth import Birth
 from betty.plugins.event_type.death import Death
 from betty.plugins.event_type.marriage import Marriage
-from betty.plugins.gender.man import Man
-from betty.plugins.gender.woman import Woman
 from betty.plugins.media_type.svg import SVG
 from betty.plugins.place_type.country import Country
 from betty.plugins.place_type.municipality import Municipality
