+++ betty/tests/plugins/asset/raspberry_mint/templates/entity/test_summary__person.py
@@ -1,3 +1,4 @@
+from betty.genders.non_binary import NonBinary
 from betty.locale.localize import DEFAULT_LOCALIZER
 from betty.plugins.asset_directory.raspberry_mint import RASPBERRY_MINT
 from betty.plugins.entity.citation import Citation
@@ -8,7 +9,6 @@ from betty.plugins.entity.presence import Presence
 from betty.plugins.entity.source import Source
 from betty.plugins.event_type.birth import Birth
 from betty.plugins.event_type.death import Death
-from betty.plugins.gender.non_binary import NonBinary
 from betty.plugins.role.subject import Subject
 from betty.privacy import Privacy
 from betty.test_utils.conftest import AssertTemplateFile
